+++ src/v1/Archetypes/Nominals/Factories/makeNominalType.ts
@@ -33,6 +33,8 @@
 //
 import { DataGuarantee, FunctionalOption } from "../../FunctionTypes";
 import { MakeNominalTypeOptions } from "./MakeNominalTypeOptions";
+import { THROW_THE_ERROR } from "../../../ErrorHandling";
+import { DEFAULT_DATA_PATH } from "../../../SupportingTypes";
 
 /**
  * `makeNominalType()` converts your input type into a branded or
@@ -45,9 +47,13 @@ import { MakeNominalTypeOptions } from "./MakeNominalTypeOptions";
  * Make sure that it has no side-effects whatsoever.
  * @param input
  * The data to brand or flavour.
+ * @param onError
+ * If `input` is rejected by the `contract`, we'll call this with an
+ * `AppError` to explain why.
+ * @param path
+ * Where are you in the data structure you're creating?
  * @param options
- * The user-supplied options object. This will be passed into `contract()`
- * and any `fnOpts`.
+ * The remaining user-supplied options. They will be passed into any `fnOpts`.
  * @param fnOpts
  * A (possibly empty) list of user-supplied functional options. These will
  * be applied once the `contract()` has successfully validated the `input`
@@ -62,21 +68,28 @@ import { MakeNominalTypeOptions } from "./MakeNominalTypeOptions";
 export function makeNominalType<IN, OUT, OPT extends object = object>(
     contract: DataGuarantee<IN, OPT>,
     input: IN,
-    options: MakeNominalTypeOptions & OPT,
+    {
+        onError = THROW_THE_ERROR,
+        path = DEFAULT_DATA_PATH,
+        ...options
+    }: Partial<MakeNominalTypeOptions> & Partial<OPT> = {},
     ...fnOpts: FunctionalOption<IN|OUT, OPT>[]
 ): OUT {
     // enforce the contract
-    contract(input, options);
+    contract(input, { onError, path, ...options });
 
     // prepare our return value
     let retval: IN|OUT = (input as unknown) as OUT;
+
+    // shorthand
+    const optsForFnOpts = { onError, path, ...options } as OPT;
 
     // apply the functional options
     //
     // we don't use `applyFunctionalOptions()` here, because
     // we need this type-cast to support IN|OUT
     fnOpts.forEach((fnOpt) => {
-        retval = (fnOpt(retval, options) as OUT);
+        retval = (fnOpt(retval, optsForFnOpts) as OUT);
     });
 
     // we have to re-cast here, because the compiler isn't sure